Add files via upload
Challenge 3: How would you introduce events into the past and future? How should someone proceed to include shocks into the extrapolation phase?

We define below what we perceive as a serious challenge to the RAI stability coin emerging as the universal or pivotal “stability coin” much like the way Bitcoin has clearly emerged as the pivotal repository for “value”.
In the context of the proportional–integral–derivative controller (PID controller), the above issue is purely in the real of the Derivative, beyond being a pure PI controller alone. Conceptually, from a research point of view, we would simulate RAI on a CEX to generate arbitrage signals from it in the form of events or price shocks and evaluate the response time of the